tests/wpt/tests/html/semantics/grouping-content/the-ul-element/grouping-ul.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the ul element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-ul-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("ul");
            assert_equals(Object.getPrototypeOf(testElement), HTMLUListElement.prototype, "HTMLUListElement.prototype should be used for ul");
        }, "The prototype for ul is HTMLUListElement.prototype");

    </script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the ul element.</p>

    <div id="log"></div>

</body>
</html>
